refactor of socket client, much more aggressive restarts on failures
